requirements.txt
altgraph==0.17
argon2-cffi==20.1.0
asgiref==3.4.1
async-generator==1.10
attrs==20.3.0
auto-py-to-exe==2.10.0
backcall==0.2.0
backports.entry-points-selectable==1.1.1
binarytree==6.3.0
bleach==3.3.0
bottle==0.12.19
bottle-websocket==0.2.9
Brotli==1.0.9
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.3
colorama==0.4.4
cycler==0.10.0
dash==2.0.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.3
Django==3.2.6
django-cors-headers==3.10.0
djangorestframework==3.12.4
Eel==0.12.4
entrypoints==0.3
et-xmlfile==1.0.1
filelock==3.4.0
Flask==2.0.2
Flask-Compress==1.10.1
future==0.18.2
gevent==21.8.0
gevent-websocket==0.10.1
graphviz==0.19
greenlet==1.1.1
idna==2.10
image==1.5.33
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jdcal==1.4.1
jedi==0.17.2
Jinja2==3.0.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
ksherpay==0.1.0
MarkupSafe==2.0.1
matplotlib==3.3.4
mistune==0.8.4
mpmath==1.2.1
mysql-connector-python==8.0.27
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
networkx==2.6.3
nose==1.3.7
notebook==6.2.0
numpy==1.20.1
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
openpyxl==3.0.5
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.3
parso==0.7.1
pefile==2021.5.24
pgzero==1.2.1
pickleshare==0.7.5
Pillow==8.1.0
platformdirs==2.4.0
plotly==5.4.0
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.19.0
psycopg2==2.9.3
pycparser==2.20
pygame==2.0.1
Pygments==2.7.2
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.2
pynput==1.7.3
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.17.1
pytube==11.0.2
pytz==2021.1
pywin32==228
pywin32-ctypes==0.2.0
pywinpty==0.5.7
pyzmq==19.0.2
qtconsole==5.0.2
QtPy==1.9.0
requests==2.25.1
robotframework==3.2.2
robotframework-pythonlibcore==2.1.0
robotframework-selenium2library==3.0.0
robotframework-seleniumlibrary==4.5.0
scipy==1.6.1
selenium==3.141.0
Send2Trash==1.5.0
setuptools-scm==6.3.2
six==1.15.0
sqlparse==0.4.1
sympy==1.7.1
tenacity==8.0.1
terminado==0.9.2
testpath==0.4.4
tomli==1.2.2
tornado==6.1
traitlets==5.0.5
trees==0.0.3
urllib3==1.26.5
virtualenv==20.10.0
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
whichcraft==0.6.1
widgetsnbextension==3.5.1
zope.event==4.5.0
zope.interface==5.4.0
